feat: updated on quit
